+++ src/Routing/DispatcherInterface.php
@@ -28,6 +28,8 @@ interface DispatcherInterface
      *
      * @param  \Rougin\Slytherin\Routing\RouterInterface $router
      * @return self
+     *
+     * @throws \UnexpectedValueException
      */
     public function setRouter(RouterInterface $router);
 }