+++ src/Routing/RouteInterface.php
@@ -38,11 +38,6 @@ interface RouteInterface
     public function getRegex();
 
     /**
-     * @return mixed
-     */
-    public function getResult();
-
-    /**
      * @return string
      */
     public function getUri();
@@ -52,10 +47,4 @@ interface RouteInterface
      * @return self
      */
     public function setParams($params);
-
-    /**
-     * @param  mixed $result
-     * @return self
-     */
-    public function setResult($result);
 }