+++ tests/Sample/Router.php
@@ -22,7 +22,7 @@ class Router extends Slytherin
 
         $this->get('/response', 'Hello@response');
 
-        $this->get('/hi/:name', 'Hello@name');
+        $this->get('/hi/{name}', 'Hello@name');
 
         $this->get('without-slash', 'Hello@string');
 
@@ -33,7 +33,12 @@ class Router extends Slytherin
             return 'Welcome call!';
         });
 
-        $this->get('/call/:name', function ($name)
+        $this->get('/call/{name}/{age}', function ($name, $age)
+        {
+            return 'Welcome ' . $name . ', ' . $age . '!';
+        });
+
+        $this->get('/call/{name}', function ($name)
         {
             return 'Welcome ' . $name . '!';
         });